// config/editor/src/app/definition/enums/PlatformType.ts
export enum PlatformType {
    DALLAS = "dallas",
    DS2482 = "ds2482",
    ANALOG = "analog",
    MODBUS = "modbus",
    LM75 = "lm75",
    I2C = "i2c",
    MCP23017 = "mcp23017",
    OLED = "oled",
    MQTT = "mqtt",
    PCA9685 = "pca9685",
    GPIO = "gpio",
    DIGITAL = "digital",
    SWITCH = "switch",
    RESISTANCE = "resistance",
    NTC = "ntc",
}